// src/app/Interface/positions.interface.ts
export interface positions {
    BaseValue: number;
    BenchmarkBenchMarkExposureColor: number;
    BenchmarkPercentOfFundNAV: number;
    Beta: number
    Book: string;
    CostPrice: number;
    CurrentPrice: number;
    CustodianValue: string;
    Delta: number;
    EntryTimeStamp: string;
    Exchange: string;
    FXRate: number;
    FundName: string;
    ISIN: string;
    ISO: string;
    InitialOpenCostBase: number;
    InitialOpenCostLocal: number;
    LastChangeDirection: string;
    LastCloseDirection: string;
    Lqdty: number;
    Lqdty_Pos_1d: number;
    MarketExposureBase: number;
    MarketValue: number;
    MarketValueLocal: number;
    //Message: string|null;
    NetPrice: number;
    NetProfitLossBase: number;
    OpeningCostBase: number;
    PercentExposureOfNAV: number;
    PercentageMovementFromLastClose: number;
    PositionClass: string;
    PositionReference: number;
    PricingCurrency: string;
    ProductDescription: string;
    ProductID: string;
    ProductType: string;
    ProfitLoss: number;
    ProfitLossBase: number;
    ProfitLossExclFX: number;
    ProfitLossFXOnly: number;
    Quantity: number;
    SEDOL: string;
    SettleCurrency: string;
    Strategy: string;
    SubPriceSource: string;
}